FIx failing test cases and rename NetworkManager
